Move fixture-anchored tests to IntegrationTesting; fetch fixtures from HF dataset
Per davidkoski's review on #308: the 43 `tools/fixtures/*.safetensors`
data files were vendored in-repo so the Rung 1–4 / mask parity tests
could run offline. The PR review flagged both the file size and the
"integration-style" placement. This commit removes the vendored data
and moves the fixture-anchored tests into the existing IntegrationTesting
Xcode project (introduced in #142, hosts integration tests that fetch
real models from HF rather than running on CI).

Moved files (entire, no edits beyond fixture rewiring):

- `BidirectionalMasksTests.swift` — 8 mask-parity tests
- `MTPRung4TokenParityTests.swift` — 3 drafter token-parity tests
- `MTPAcceptanceRateTests.swift` — 1 placeholder (full impl deferred)
- `MTPQuantizationOnsetTests.swift` — 1 placeholder (full impl deferred)

Each moved file now:

- Imports `IntegrationTestHelpers` and uses the `downloadDataset(...)`
  helper added in the previous commit.
- Pins the fixture revision to
  `ae79d3536d023d1fd0db4fa829c1d29adadf7bd1` (the "Initial MTP fixtures,
  mlx-vlm@d49d428, target=31b-8bit, drafter=31B-bf16" commit on
  https://huggingface.co/datasets/angelsbrood/gemma4-mtp-fixtures).
- Preserves the existing `Issue.record(...)` skip-with-warning pattern
  for cases where the dataset or checkpoints aren't reachable.

Deleted: all 43 `tools/fixtures/{masks,drafter_forward,drafter_block,
end_to_end}/*.safetensors` files and the four now-empty subdirectories.

Kept: `tools/fixtures/FIXTURE-SCHEMA.md`, `FIXTURE-MANIFEST.json`,
`tools/generate_mtp_fixtures.py`, `tools/inspect_drafter_layout.py`,
plus a new `tools/fixtures/README.md` pointing readers to the HF
dataset and the pinned SHA. Schema, manifest, and generation tooling
stay in-repo for reproducibility.

`Gemma4AssistantDraftModelTests.swift` and `MTPDrafterModelFactoryTests
.swift` mix synthetic-mock and checkpoint/fixture-gated cases; they are
split in the next commit so the synthetic half stays in `MLXLMTests`
for CI coverage.

Verified locally: `swift build` clean, `xcodebuild build-for-testing
-scheme IntegrationTesting` succeeds; the moved tests compile against
the IntegrationTesting target.
