+++ IntegrationTesting/IntegrationTestingTests/MTPAcceptanceRateTests.swift
@@ -1,0 +1,61 @@
+// Copyright © 2026 Apple Inc.
+
+import Foundation
+import MLX
+import MLXLMCommon
+import MLXVLM
+import Testing
+
+// MARK: - Helpers
+
+private func hfSnapshotDir(modelId: String) -> URL? {
+    let home = FileManager.default.homeDirectoryForCurrentUser
+    let hub = home.appendingPathComponent(".cache/huggingface/hub")
+    let folderName = "models--" + modelId.replacingOccurrences(of: "/", with: "--")
+    let snapshots = hub.appendingPathComponent(folderName).appendingPathComponent("snapshots")
+    guard
+        let entries = try? FileManager.default.contentsOfDirectory(
+            at: snapshots, includingPropertiesForKeys: nil)
+    else { return nil }
+    return entries.first
+}
+
+// MARK: - Acceptance-rate floor
+
+/// PLAN §11 / R12 sets a loose floor of `>= 0.30` for the 64-token,
+/// blockSize=4, temperature=0 production-correctness gate. mlx-vlm reports
+/// a 3.94x speedup on the same case which implies acceptance closer to
+/// 0.7-0.8; the floor leaves headroom for hardware variance and
+/// non-determinism in SDPA kernel ordering.
+///
+/// Tests are gated by the presence of both target and drafter checkpoints
+/// in the HF cache.
+
+@Test
+func testAcceptanceRateFloor64TokenBlock4Temp0() async throws {
+    // Documents the verification plan. Full target+drafter end-to-end
+    // measurement is deferred to a follow-up PR; this placeholder gates on
+    // checkpoint presence so the test reports a useful skip reason when the
+    // moved IntegrationTesting target runs without the prerequisite caches.
+    guard hfSnapshotDir(modelId: "mlx-community/gemma-4-31B-it-assistant-bf16") != nil
+    else {
+        Issue.record("31B-assistant-bf16 not in HF cache; skipping acceptance-rate floor")
+        return
+    }
+    guard hfSnapshotDir(modelId: "mlx-community/gemma-4-31b-it-8bit") != nil
+    else {
+        Issue.record("31b-it-8bit target not in HF cache; skipping acceptance-rate floor")
+        return
+    }
+
+    Issue.record(
+        """
+        acceptance-rate floor test is placeholdered: full target+drafter
+        end-to-end measurement (64 tokens, blockSize=4, temperature=0,
+        floor ≥ 0.30 per PLAN §11/R12) is deferred to a follow-up PR.
+        The Rung 4 token-parity tests in this target cover the drafter's
+        `draftBlock(...)` directly; this test will be wired up once the
+        target+iterator loop is exercisable here.
+        """
+    )
+}
