--- Tests/MLXLMTests/MTPAcceptanceRateTests.swift
@@ -1,68 +1,0 @@
-// Copyright © 2026 Apple Inc.
-
-import Foundation
-import MLX
-import MLXLMCommon
-import MLXVLM
-import Testing
-
-// MARK: - Helpers
-
-private func hfSnapshotDir(modelId: String) -> URL? {
-    let home = FileManager.default.homeDirectoryForCurrentUser
-    let hub = home.appendingPathComponent(".cache/huggingface/hub")
-    let folderName = "models--" + modelId.replacingOccurrences(of: "/", with: "--")
-    let snapshots = hub.appendingPathComponent(folderName).appendingPathComponent("snapshots")
-    guard
-        let entries = try? FileManager.default.contentsOfDirectory(
-            at: snapshots, includingPropertiesForKeys: nil)
-    else { return nil }
-    return entries.first
-}
-
-// MARK: - Acceptance-rate floor
-
-/// PLAN §11 / R12 sets a loose floor of `>= 0.30` for the 64-token,
-/// blockSize=4, temperature=0 production-correctness gate. mlx-vlm reports
-/// a 3.94x speedup on the same case which implies acceptance closer to
-/// 0.7-0.8; the floor leaves headroom for hardware variance and
-/// non-determinism in SDPA kernel ordering.
-///
-/// Tests are gated by the presence of both target and drafter checkpoints
-/// in the HF cache.
-
-@Test
-func testAcceptanceRateFloor64TokenBlock4Temp0() async throws {
-    // Documents the verification plan. The test gates on checkpoint
-    // presence AND on a runtime that can load both the 31b-it-8bit target
-    // via `VLMModelFactory` and the 31B-it-assistant-bf16 drafter via
-    // `MTPDrafterModelFactory`. The target load requires a real tokenizer
-    // loader (`#huggingFaceTokenizerLoader()` from MLXHuggingFace, which
-    // is not in MLXLMTests' dependency closure), so the actual end-to-end
-    // path runs in Xcode with the full app environment rather than in
-    // `swift test`. The test is structured so that when both pieces are
-    // available, it exercises the path; otherwise it skips with a clear
-    // record.
-    guard hfSnapshotDir(modelId: "mlx-community/gemma-4-31B-it-assistant-bf16") != nil
-    else {
-        Issue.record("31B-assistant-bf16 not in HF cache; skipping acceptance-rate floor")
-        return
-    }
-    guard hfSnapshotDir(modelId: "mlx-community/gemma-4-31b-it-8bit") != nil
-    else {
-        Issue.record("31b-it-8bit target not in HF cache; skipping acceptance-rate floor")
-        return
-    }
-
-    Issue.record(
-        """
-        acceptance-rate floor test is intentionally skipped: target loading
-        requires `#huggingFaceTokenizerLoader()` from MLXHuggingFace, which
-        is not currently a dependency of MLXLMTests. The verification path
-        runs against `Gemma4AssistantDraftModel.draftBlock(...)` directly
-        in `MTPRung4TokenParityTests` (Rung 4); full target+drafter
-        acceptance-rate measurement runs in a downstream app/integration
-        test once the tokenizer loader is wired in.
-        """
-    )
-}
